requirements.txt
aiohttp==3.6.2
APScheduler==3.6.3
arrow==0.15.5
async-timeout==3.0.1
attrs==19.3.0
base58==2.0.0
beautifulsoup4==4.8.2
blinker==1.4
callofduty.py==1.2.1
certifi==2020.6.20
chardet==3.0.4
click==6.6
cssutils==1.0.2
databases==0.2.6
dj-config-url==0.1.1
dnspython==1.16.0
docker-py==1.8.0
ecdsa==0.15
emoji==0.5.4
entrypoints==0.3
eventlet==0.28.0
facebook-sdk==3.1.0
facebookads==2.11.4
flake8==3.7.9
Flask==1.1.1
Flask-Mail==0.9.1
Flask-SocketIO==4.2.1
Flask-WTF==0.14.2
funcsigs==1.0.2
future==0.18.2
greenlet==0.4.17
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2020.9.29
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
jsonrpcclient==2.0.1
jsonrpcserver==3.1.1
jsonschema==3.2.0
lxml==4.4.2
MarkupSafe==1.1.1
mccabe==0.6.1
mnemonic==0.13
monotonic==1.5
more-itertools==8.2.0
multidict==4.7.4
numpy==1.18.1
opencv-python==4.2.0.34
packaging==20.1
pandas==0.25.3
passlib==1.7.2
path==13.1.0
path.py==12.4.0
pbkdf2==1.3
pexpect==4.8.0
pluggy==0.13.1
protobuf==3.0.0a3
ptyprocess==0.6.0
py==1.8.1
pyaes==1.6.1
pycodestyle==2.5.0
pycryptodome==3.9.6
pyflakes==2.1.1
pygal==2.4.0
PyJWT==1.7.1
PyMySQL==0.9.3
PyMysqlDB==0.0.2
pyparsing==2.4.6
pyrsistent==0.15.7
pytest==5.3.5
python-dateutil==2.8.1
python-engineio==3.11.2
python-socketio==4.4.0
python-twitch-client==0.6.0
pytz==2019.3
pywallet==0.1.0
PyYAML==5.3
requests==2.23.0
rfc3986==1.4.0
schedule==0.6.0
selenium==3.141.0
sha256==0.1
six==1.14.0
sniffio==1.1.0
soupsieve==1.9.5
SQLAlchemy==1.3.13
stripe==2.47.0
tabulate==0.8.6
telepot==12.7
twilio==6.39.0
twitchAPI==1.3
two1==3.10.9
tzlocal==2.0.0
urljoin==1.0.0
urllib3==1.24.1
urlopen==1.0.0
virtualenv==16.7.9
wcwidth==0.1.8
websocket-client==0.57.0
Werkzeug==0.16.0
wikipedia==1.4.0
WTForms==2.2.1
yarl==1.4.2
zope.event==4.5.0
zope.interface==5.1.2
